+++ src/main/java/com/zepben/vertxutils/routing/chunked/HttpChunkedJsonResponse.kt
@@ -7,6 +7,7 @@
  */
 package com.zepben.vertxutils.routing.chunked
 
+import io.netty.handler.codec.http.HttpResponseStatus
 import io.vertx.core.http.HttpServerResponse
 
 class HttpChunkedJsonResponse(
@@ -14,17 +15,29 @@ class HttpChunkedJsonResponse(
     private val bufferSize: Int = DEFAULT_BUFFER_SIZE,
 ) : ChunkedJsonResponse(bufferSize) {
 
+    private var canSetStatus = true
+
     override fun onResponseCompleted(sb: StringBuilder) {
+        canSetStatus = false
         if (!response.closed()) {
             response.end(sb.toString())
         }
     }
 
     override fun checkWrite(sb: StringBuilder, force: Boolean) {
+        canSetStatus = false
         if ((force || (sb.length >= bufferSize)) && !response.closed()) {
             response.write(sb.toString())
             sb.setLength(0)
         }
     }
+
+    var statusCode: HttpResponseStatus
+        get() = HttpResponseStatus.valueOf(response.statusCode)
+        set(status) {
+            // Once a response is committed (first write or end), it is too late to change the status.
+            check(canSetStatus) { "You can't set the status after the response has been committed" }
+            response.statusCode = status.code()
+        }
 
 }